requirements.txt
streamlit==1.24.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
plotly==5.15.0
pytz==2023.3
python-dateutil==2.8.2
